event broker prunes waits when their context gets cancelled
The number of waits tracked by the broker was growing unbounded. Waits
would only get removed when an event was sent to them, so rare events
would pile up.

This change adds a unique ID to each context and enables the broker to
track which waits correspond to a specific context, and remove them when
that context gets cancelled.